refactor(api): extraer configureApp compartido entre main.ts y E2E
Extrae helmet/compression/CORS/prefix/versioning/Swagger a un helper
configureApp() en app.config.ts; tanto main.ts como app.e2e-spec.ts lo
invocan, eliminando la duplicación que causaba drift silencioso. Aplica
también review note M1: agrega message: expect.any(String) al matcher
del test 404.
