+++ apps/api/src/main.ts
@@ -1,11 +1,8 @@
 import 'reflect-metadata';
-import { VersioningType } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import compression from 'compression';
-import helmet from 'helmet';
 import { Logger } from 'nestjs-pino';
+import { configureApp } from './app.config';
 import { AppModule } from './app.module';
 import type { Env } from './config/env.schema';
 
@@ -16,40 +13,7 @@ async function bootstrap() {
   app.enableShutdownHooks();
 
   const config = app.get<ConfigService<Env, true>>(ConfigService);
-  const swaggerEnabled = config.get('SWAGGER_ENABLED', { infer: true });
-
-  app.use(
-    helmet({
-      // Swagger UI requires inline scripts/styles that helmet's default CSP blocks.
-      // Safe to disable globally: this is a JSON API with no HTML routes outside /api/docs.
-      contentSecurityPolicy: swaggerEnabled ? false : undefined,
-    }),
-  );
-  app.use(compression());
-
-  app.enableCors({
-    origin: config.get('CORS_ORIGINS', { infer: true }),
-    credentials: true,
-  });
-
-  app.setGlobalPrefix('api', { exclude: ['/health'] });
-  app.enableVersioning({
-    type: VersioningType.URI,
-    defaultVersion: '1',
-  });
-
-  if (swaggerEnabled) {
-    const swaggerConfig = new DocumentBuilder()
-      .setTitle('Gestión Académica API')
-      .setDescription('API del sistema de gestión académica')
-      .setVersion('1.0')
-      .addBearerAuth()
-      .build();
-    const document = SwaggerModule.createDocument(app, swaggerConfig);
-    SwaggerModule.setup('api/docs', app, document, {
-      jsonDocumentUrl: 'api/docs-json',
-    });
-  }
+  configureApp(app, config);
 
   const port = config.get('PORT', { infer: true });
   await app.listen(port);
